Moving private methods into extensions
