+++ Habit Helper/Controllers/Create Animal/Cell/AnimalCell.swift
@@ -1,10 +1,3 @@
-//
-//  AnimalCell.swift
-//  Habit Helper
-//
-//  Created by Анастасия Гаранович on 02.09.2021.
-//
-
 import UIKit
 
 class AnimalCell: UICollectionViewCell {
