+++ Habit Helper/Controllers/Create Habit/Cell/ColorCell.swift
@@ -1,10 +1,3 @@
-//
-//  ColorCell.swift
-//  Habit Helper
-//
-//  Created by Анастасия Гаранович on 27.07.2021.
-//
-
 import UIKit
 
 class ColorCell: UICollectionViewCell {
