+++ Habit Helper/Model/DateSelection.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+class DateSelection {
+    var date: Date
+    var selection: Bool
+    
+    init(_ date: Date) {
+        self.date = date
+        selection = false
+    }
+}
